+++ core/primitives/src/types.rs
@@ -9,7 +9,6 @@ use crate::challenge::ChallengesResult;
 use crate::hash::CryptoHash;
 use crate::serialize::u128_dec_format;
 use crate::trie_key::TrieKey;
-use std::ops::Mul;
 
 /// Account identifier. Provides access to user's state.
 pub type AccountId = String;
@@ -514,44 +513,3 @@ pub enum ValidatorKickoutReason {
 
 #[derive(PartialEq, Eq, Clone, Debug, BorshSerialize, BorshDeserialize, Serialize)]
 pub struct StateHeaderKey(pub ShardId, pub CryptoHash);
-
-/// Represents a rational number. This struct is used exclusively in configs so
-/// `u32` provides enough precision. Since we do not expect to have computation
-/// that returns a `Fraction`, there is no guarantee that it is an irreducible fraction.
-#[derive(Debug, Serialize, Deserialize, Hash, PartialEq, Eq, Clone, Copy)]
-pub struct Fraction {
-    pub numerator: u32,
-    pub denominator: u32,
-}
-
-impl Fraction {
-    pub fn new(numerator: u32, denominator: u32) -> Self {
-        Fraction { numerator, denominator }
-    }
-
-    pub fn zero() -> Self {
-        Fraction { numerator: 0, denominator: 1 }
-    }
-}
-
-impl Default for Fraction {
-    fn default() -> Self {
-        Self::zero()
-    }
-}
-
-impl Mul<u128> for Fraction {
-    type Output = u128;
-
-    fn mul(self, rhs: u128) -> Self::Output {
-        rhs * u128::from(self.numerator) / u128::from(self.denominator)
-    }
-}
-
-impl Mul<Fraction> for u128 {
-    type Output = u128;
-
-    fn mul(self, rhs: Fraction) -> Self::Output {
-        rhs * self
-    }
-}
