+++ test-utils/testlib/src/node/mod.rs
@@ -7,10 +7,11 @@ use std::sync::RwLock;
 use near_chain_configs::Genesis;
 use near_crypto::{InMemorySigner, Signer};
 use near_jsonrpc::ServerError;
+use near_primitives::fraction::Fraction;
 use near_primitives::serialize::to_base64;
 use near_primitives::state_record::StateRecord;
 use near_primitives::transaction::SignedTransaction;
-use near_primitives::types::{AccountId, Balance, Fraction, NumSeats};
+use near_primitives::types::{AccountId, Balance, NumSeats};
 use near_primitives::validator_signer::InMemoryValidatorSigner;
 use near_primitives::views::AccountView;
 use neard::config::{
